Created debug assets and scripts for the game mode
All assets are debug objects (using black and purple squares). The
objects have all scripts necessary to function, all objects are relative
to the deck object so should allow for a smooth transition to an AR
camera.
Next steps: Change the assets to use the graphics with the rest of the
project, use an AR camera with the game mode, implement a functional
score system (the current script adds a score to a public variable and
needs to be changed to use the player prefs from the behind the scenes
score system being made by Adam).